refactor(reviews): split cutover review modules
